Move test for Token::from_string to the gcloud auth'd user because they are tightly coupled
